pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "stompbox"
version = "1.0.0"
description = "MIDI-controlled audio pedalboard with TUI"
requires-python = ">=3.11"
dependencies = [
    "pedalboard>=0.9",
    "textual>=0.80",
    "sounddevice>=0.4.6",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "mido>=1.3",
]

[project.scripts]
stompbox = "stompbox.cli:main"

[tool.setuptools.packages.find]
where = ["."]
